Initial implementation of fast failovers
This patch tries to connect to a database configured as a writer in the
`databases` section.

If DNS resolution fails, then a connection is made
to the configured reader. The `topology_query` is used to get the
topology information from the reader.

If the reader's DNS hostname matches the configured writer, then nothing
else needs to be done and the connection resumes. Otherwise, the
connection is closed and a new connection is opened to the second node
in the cluster which is the updated writer.

This solution should scale to N nodes as long as replication lag is at a
minimum and the topology information provides who the new writer is.

The query used to get topology was the following:

```
select * from topology where is_writer = 't';
```

The table created ahead of time is the following:

```
postgres=> select * from topology;
                                    hostname                                     | is_writer
---------------------------------------------------------------------------------+-----------
 reader.baz.com    | t
 reader-2.baz.com    | f
(2 rows)
```

This patch also adds support for escaping a single quote in the connect query for a
database config.

Example:

```
topology_query='select * from topology where is_writer = \'t\';'
```