Do not launch new connections unless a global writer was detected
This can have the side effect of trying to bring back up a failed node
while a switchover is taking place. If `launch_new_connection` is called
while that happens, the DNS lookup and associated connection attempt can
add 15-30 seconds in the switchover.

Once global_writer is set, it is safe to assume that switchovers have
completed successfully and writes to the writer are no longer blocked.