Global writer should be mapped to one DB writer
Before, the global writer was global to the entire PgBouncer
configuration. This caused issues if a DB was configured as a writer
(with a topology_query) and a DB was not using the fast_switchover
feature. To solve this, have a mapping of parent nodes and a global
writer only if the DB is using fast switchovers.