Add support for recreating connection pools automatically
If a switchover takes place, and all of the nodes in the cluster have
disconnected for some reason (timeout or the cluster is down because of
a sev2), PgBouncer would loop forever and not attempt to create new
connections because new connections require DNS resolution (slow).

Now, if no nodes in the cluster are available for connection, and the
new config option `recreate_disconnected_pools` is set, try to recreate
connection pools to the available nodes.

- Add documentation about `topology_query` that was missing
- Add documentation explaining new `recreate_disconnected_pools` setting